requirements.txt
Flask
python-dotenv
requests
Pillow
google-generativeai
beautifulsoup4
webdriver-manager
selenium
python-pptx
